refactor: remove redundant directories and express module from dependencies
